+++ packages/models/src/box/sensorLayouts/sensorDefinitions/sps30_pm4.js
@@ -1,0 +1,8 @@
+'use strict';
+
+module.exports = {
+  title: 'PM4',
+  unit: 'µg/m³',
+  sensorType: 'SPS30',
+  icon: 'osem-cloud'
+};