+++ packages/models/CHANGELOG.md
@@ -20,6 +20,10 @@
 - Added `findLatestMeasurementsForSensorsWithCount` static method to Measurement schema (#588)
 - Added `sharedBoxes` functionality (#605)
 
+## v1.3.1
+- Add SPS30 sensor 
+- Update @sensebox/node-sketch-templater to v1.13.0
+
 ## v1.2.0
 
 - Extend `luftdatenHandler` (#578)